+++ tests/testthat/test-range-dataset.R
@@ -16,3 +16,13 @@ test_succeeds("range_dataset creates a dataset", {
   expect_equal(res, array(1L:10L))
 
 })
+
+
+test_succeeds("random_integer_dataset creates a dataset", {
+    ds1 <- random_integer_dataset(seed=4L) %>% dataset_take(10)
+    ds2 <- random_integer_dataset(seed=4L) %>% dataset_take(10)
+    # TODO: reticulate::iterate simplify doesn't work on objects with class(x)=="numeric"
+    r1 <- reticulate::iterate(ds1, as.numeric) %>% unlist()
+    r2 <- reticulate::iterate(ds2, as.numeric) %>% unlist()
+    expect_equal(r1, r2)
+})