+++ NEWS.md
@@ -1,3 +1,5 @@
+# tfdatasets (development version)
+
 # tfdatasets 2.4.0
 
 - Fixed problem when saving models with feature specs (#82).