+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Added `dataset_options()` for setting and getting dataset options.
 - Added a `length()` method for tensorflow datasets.
+- Added `dataset_enumerate()`
 
 # tfdatasets 2.4.0