+++ NEWS.md
@@ -4,6 +4,8 @@
 - Added a `length()` method for tensorflow datasets.
 - Added `dataset_enumerate()`
 - Added `random_integer_dataset()`
+- Added `dataset_scan()`, a stateful varient of `dataset_map()`
+
 
 # tfdatasets 2.4.0